bill-splitter-app/requirements.txt
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
filestack-python==3.2.1
Flask==1.1.2
fpdf==1.7.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pillow==10.2.0
reportlab==4.0.9
requests==2.23.0
trafaret==2.0.2
urllib3==1.25.11
Werkzeug==1.0.1
WTForms==2.3.3
